web_demo/requirements.txt
# Web Demo Requirements
gradio>=4.0.0
pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
